#14 fix data type

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	int x;
-	int y;
+	unsigned long x;
+	unsigned long y;
 	int n;
 	unsigned long sum;
